# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Filmora - Movie indexing and streaming platform powered by Telegram bot
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Movie {
  id: string;
  title: string;
  /** @nullable */
  year?: number | null;
  genre?: string[];
  /** @nullable */
  language?: string | null;
  /** @nullable */
  quality?: string | null;
  /** @nullable */
  poster?: string | null;
  /** @nullable */
  backdrop?: string | null;
  /** @nullable */
  overview?: string | null;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  imdbId?: string | null;
  /** @nullable */
  fileId?: string | null;
  /** @nullable */
  fileSize?: number | null;
  /** @nullable */
  caption?: string | null;
  addedAt: string;
}

export interface MovieList {
  items: Movie[];
  total: number;
  page: number;
  pages: number;
}

export interface FeaturedMovies {
  latest: Movie[];
  trending: Movie[];
  topRated: Movie[];
}

export interface GenreCount {
  name: string;
  count: number;
}

export interface LanguageCount {
  name: string;
  count: number;
}

export interface MovieStats {
  total: number;
  genres: GenreCount[];
  languages: LanguageCount[];
}

export interface Genre {
  name: string;
  count: number;
}

export interface ErrorResponse {
  error: string;
}

export type ListMoviesParams = {
/**
 * Search query
 */
q?: string;
/**
 * Filter by genre
 */
genre?: string;
/**
 * Filter by language
 */
language?: string;
/**
 * Filter by quality (e.g. 1080p, 720p, 480p)
 */
quality?: string;
/**
 * Page number (default 1)
 */
page?: number;
/**
 * Results per page (default 20)
 */
limit?: number;
};