# lib/api-zod/src/generated/types/errorResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Filmora - Movie indexing and streaming platform powered by Telegram bot
 * OpenAPI spec version: 0.1.0
 */

export interface ErrorResponse {
  error: string;
}