# lib/api-zod/src/generated/types/movie.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Filmora - Movie indexing and streaming platform powered by Telegram bot
 * OpenAPI spec version: 0.1.0
 */

export interface Movie {
  id: string;
  title: string;
  /** @nullable */
  year?: number | null;
  genre?: string[];
  /** @nullable */
  language?: string | null;
  /** @nullable */
  quality?: string | null;
  /** @nullable */
  poster?: string | null;
  /** @nullable */
  backdrop?: string | null;
  /** @nullable */
  overview?: string | null;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  imdbId?: string | null;
  /** @nullable */
  fileId?: string | null;
  /** @nullable */
  fileSize?: number | null;
  /** @nullable */
  caption?: string | null;
  addedAt: string;
}