# lib/api-zod/src/generated/types/movieList.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Filmora - Movie indexing and streaming platform powered by Telegram bot
 * OpenAPI spec version: 0.1.0
 */
import type { Movie } from './movie';

export interface MovieList {
  items: Movie[];
  total: number;
  page: number;
  pages: number;
}